Copying files from whiteboard-transport
The original plan was to make whiteboard-transport a submodule of both
whiteboard-mobile and the viewer app (whiteboard). However, this was a
pain for development. Since the viewer is more actively used, the code
now lives here, and mobile can be a submodule of viewer to share the
JavaScript implementation of transport.